+++ lib/help_desk/help_desk_reply/view/widget/desktop/description_desktop.dart
