+++ lib/help_desk/help_desk_reply/view/widget/desktop/help_desk_reply_desktop.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:senior_project/core/template_desktop/view/page/template_desktop.dart';
-import 'package:senior_project/help_desk/help_desk_reply/view/desktop/widget/body_reply_desktop.dart';
+import 'package:senior_project/help_desk/help_desk_reply/view/widget/desktop/body_reply_desktop.dart';
 
 class HelpDeskReplyDesktop extends StatelessWidget {
   const HelpDeskReplyDesktop({super.key});
