+++ lib/help_desk/help_desk_reply/view/widget/mobile/description_mobile.dart
